feat: handle関数をorganisms/CheckBoxContainerに移動
