+++ components/atoms/CheckBox/index.tsx
@@ -5,6 +5,7 @@ interface Props {
   label: string;
   prefCode: number;
   prefName: string;
+  onChange: (prefCode: number, prefName: string, check: boolean) => void;
 }
 
 const Styles: { [key: string]: CSSProperties } = {
@@ -12,21 +13,18 @@ const Styles: { [key: string]: CSSProperties } = {
   label: {},
 };
 
-const CheckBox: VFC<Props> = ({ id, label, prefCode, prefName }: Props) => {
-  const handleOnChangeCheckBox = (
-    prefCode: number,
-    prefName: string,
-    check: boolean
-  ) => {
-    console.log(prefCode, prefName, check);
-  };
+const CheckBox: VFC<Props> = ({
+  id,
+  label,
+  prefCode,
+  prefName,
+  onChange,
+}: Props) => {
   return (
     <div style={Styles.checkBox}>
       <input
         id={id}
-        onChange={(e) =>
-          handleOnChangeCheckBox(prefCode, prefName, e.target.checked)
-        }
+        onChange={(e) => onChange(prefCode, prefName, e.target.checked)}
         type="checkbox"
       />
       <label htmlFor={id} style={Styles.label}>
